+++ KsApi/Service+RequestHelpers.swift
@@ -36,63 +36,6 @@ extension Service {
     }
   }
 
-  func requestPagination<M: Decodable>(_ paginationUrl: String)
-    -> SignalProducer<M, ErrorEnvelope> where M == M.DecodedType {
-    guard let paginationUrl = URL(string: paginationUrl) else {
-      return .init(error: .invalidPaginationUrl)
-    }
-
-    return Service.session.rac_JSONResponse(preparedRequest(forURL: paginationUrl))
-      .flatMap(self.decodeModel)
-  }
-
-  func request<M: Decodable>(_ route: Route)
-    -> SignalProducer<M, ErrorEnvelope> where M == M.DecodedType {
-    let properties = route.requestProperties
-
-    guard let URL = URL(string: properties.path, relativeTo: self.serverConfig.apiBaseUrl as URL) else {
-      fatalError(
-        "URL(string: \(properties.path), relativeToURL: \(self.serverConfig.apiBaseUrl)) == nil"
-      )
-    }
-
-    return Service.session.rac_JSONResponse(
-      preparedRequest(forURL: URL, method: properties.method, query: properties.query),
-      uploading: properties.file.map { ($1, $0.rawValue) }
-    )
-    .flatMap(self.decodeModel)
-  }
-
-  func request<M: Decodable>(_ route: Route)
-    -> SignalProducer<[M], ErrorEnvelope> where M == M.DecodedType {
-    let properties = route.requestProperties
-
-    let url = self.serverConfig.apiBaseUrl.appendingPathComponent(properties.path)
-
-    return Service.session.rac_JSONResponse(
-      preparedRequest(forURL: url, method: properties.method, query: properties.query),
-      uploading: properties.file.map { ($1, $0.rawValue) }
-    )
-    .flatMap(self.decodeModels)
-  }
-
-  func request<M: Decodable>(_ route: Route)
-    -> SignalProducer<M?, ErrorEnvelope> where M == M.DecodedType {
-    let properties = route.requestProperties
-
-    guard let URL = URL(string: properties.path, relativeTo: self.serverConfig.apiBaseUrl as URL) else {
-      fatalError(
-        "URL(string: \(properties.path), relativeToURL: \(self.serverConfig.apiBaseUrl)) == nil"
-      )
-    }
-
-    return Service.session.rac_JSONResponse(
-      preparedRequest(forURL: URL, method: properties.method, query: properties.query),
-      uploading: properties.file.map { ($1, $0.rawValue) }
-    )
-    .flatMap(self.decodeModel)
-  }
-
   func requestDecodable<M: Swift.Decodable>(_ route: Route)
     -> SignalProducer<M, ErrorEnvelope> {
     let properties = route.requestProperties
