+++ KsApi/ServiceTests.swift
@@ -3,6 +3,14 @@ import ReactiveExtensions
 import ReactiveSwift
 import XCTest
 
+struct MySwiftModel: Swift.Decodable, Equatable {
+  public let array: [String]
+  public let bool: Bool
+  public let dict: [String: String]
+  public let id: Int
+  public let name: String
+}
+
 final class ServiceTests: XCTestCase {
   func testDefaults() {
     XCTAssertTrue(Service().serverConfig == ServerConfig.production)
@@ -43,90 +51,6 @@ final class ServiceTests: XCTestCase {
     let loggedOut = loggedIn.logout()
 
     XCTAssertTrue(loggedOut == Service())
-  }
-
-  func testArgoDecodeModel_ValidModel() {
-    let data: [String: Any] = [
-      "id": 1,
-      "name": "Argo Name",
-      "model": [
-        "array": ["string1", "string2"],
-        "bool": true,
-        "dict": ["key1": "value1", "key2": "value2"],
-        "id": 5,
-        "name": "Swift Name"
-      ]
-    ]
-
-    let expectedResult = MyArgoModel.decodeJSONDictionary(data).value
-
-    let service = Service()
-    let model: SignalProducer<MyArgoModel, ErrorEnvelope> = service.decodeModel(data)
-    let result = try! model.single()?.get()
-    XCTAssertEqual(result, expectedResult)
-  }
-
-  func testArgoDecodeModel_WrongKeyModel() {
-    let data: [String: Any] = [
-      "id": 1,
-      "name": "Argo Name",
-      "model": [
-        "array": ["string1", "string2"],
-        "bool": true,
-        "dict": ["key1": "value1", "key2": "value2"],
-        "id": 5,
-        "wrong_key": "Swift Name"
-      ]
-    ]
-
-    let model: SignalProducer<MyArgoModel, ErrorEnvelope> = Service().decodeModel(data)
-    XCTAssertThrowsError(try model.single()?.get(), "wrong key should throw an error")
-  }
-
-  func testArgoDecodeModel_WrongTypeModel() {
-    let data: [String: Any] = [
-      "id": 1,
-      "name": "Argo Name",
-      "model": [
-        "array": ["string1", "string2"],
-        "bool": "wrong type",
-        "dict": ["key1": "value1", "key2": "value2"],
-        "id": 5,
-        "name": "Swift Name"
-      ]
-    ]
-
-    let model: SignalProducer<MyArgoModel, ErrorEnvelope> = Service().decodeModel(data)
-    XCTAssertThrowsError(try model.single()?.get(), "wrong type should throw an error")
-  }
-
-  func test_Array_ArgoDecodeModel_ValidModel() {
-    let data: [[String: Any]] = [[
-      "id": 1,
-      "name": "Argo Name",
-      "model": [
-        "array": ["string1", "string2"],
-        "bool": true,
-        "dict": ["key1": "value1", "key2": "value2"],
-        "id": 5,
-        "name": "Swift Name"
-      ]
-    ]]
-
-    let expectedResult = data.map { (dict) -> MyArgoModel in
-      MyArgoModel.decodeJSONDictionary(dict).value!
-    }
-
-    let model: SignalProducer<[MyArgoModel], ErrorEnvelope> = Service().decodeModels(data)
-    let result = try! model.single()?.get()
-    XCTAssertEqual(result, expectedResult)
-  }
-
-  func test_Array_ArgoDecodeModel_EmptyArray() {
-    let data: [[String: Any]] = [[:]]
-
-    let model: SignalProducer<[MyArgoModel], ErrorEnvelope> = Service().decodeModels(data)
-    XCTAssertThrowsError(try model.single()?.get(), "empty array should throw an error")
   }
 
   func testSwiftDecodeModel_ValidModel() {
