+++ KsApi/models/ShippingRule.swift
@@ -27,7 +27,3 @@ public func == (lhs: ShippingRule, rhs: ShippingRule) -> Bool {
   // TODO: change to compare id once that api is deployed
   return lhs.location == rhs.location
 }
-
-private func stringToDouble(_ string: String) -> Decoded<Double> {
-  return Double(string).map(Decoded.success) ?? .success(0)
-}
