--- Library/ArgoDecoded/Argo.swift
@@ -1,12 +1,0 @@
-import Runes
-
-precedencegroup ArgoDecodePrecedence {
-  associativity: left
-  higherThan: RunesApplicativeSequencePrecedence
-  lowerThan: NilCoalescingPrecedence
-}
-
-infix operator <|: ArgoDecodePrecedence
-infix operator <|?: ArgoDecodePrecedence
-infix operator <||: ArgoDecodePrecedence
-infix operator <||?: ArgoDecodePrecedence
