--- Library/ArgoDecoded/NSNumber.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-extension NSNumber {
-  var isBool: Bool {
-    return type(of: self) == type(of: NSNumber(value: true))
-  }
-}
